Add files via upload
Here is a "case switch" file with both of our games - it needs all the graphics and sounds from both games to be in the same folder as the CaseSwitch to work, but I have tested it and it works on my Mac. I added a bit of text to the start of your game to instruct the player to use the arrow keys to move the yellow car.